fix(go/genkit): Fix index function
